+++ src/main/java/org/torproject/metrics/onionoo/docs/DateTimeHelper.java
@@ -11,7 +11,6 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.TimeZone;
 
 public class DateTimeHelper {
 
@@ -80,7 +79,6 @@ public class DateTimeHelper {
     if (!threadDateFormats.containsKey(format)) {
       DateFormat dateFormat = new SimpleDateFormat(format);
       dateFormat.setLenient(false);
-      dateFormat.setTimeZone(TimeZone.getTimeZone("UTC"));
       threadDateFormats.put(format, dateFormat);
     }
     return threadDateFormats.get(format);