+++ src/main/java/org/torproject/metrics/onionoo/server/ServerMain.java
@@ -9,6 +9,9 @@ import org.eclipse.jetty.xml.XmlConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.Locale;
+import java.util.TimeZone;
+
 public class ServerMain {
 
   private static final Logger logger = LoggerFactory.getLogger(
@@ -16,6 +19,8 @@ public class ServerMain {
 
   /** Starts the web server listening for incoming client connections. */
   public static void main(String[] args) {
+    Locale.setDefault(Locale.US);
+    TimeZone.setDefault(TimeZone.getTimeZone("UTC"));
     try {
       Resource onionooXml = Resource.newSystemResource("jetty.xml");
       logger.info("Reading configuration from '{}'.", onionooXml);